'Merge' sound codes from 'https://source.codeaurora.org/quic/la/kernel/msm-4.4/tree/?h=kernel.lnx.4.4.r38-rel' branch

+++ sound/usb/mixer_quirks.c
@@ -755,7 +755,7 @@ static int snd_ni_control_init_val(struct usb_mixer_interface *mixer,
 		return err;
 	}
 
-	kctl->private_value |= ((unsigned int)value << 24);
+	kctl->private_value |= (value << 24);
 	return 0;
 }
 
@@ -916,7 +916,7 @@ static int snd_ftu_eff_switch_init(struct usb_mixer_interface *mixer,
 	if (err < 0)
 		return err;
 
-	kctl->private_value |= (unsigned int)value[0] << 24;
+	kctl->private_value |= value[0] << 24;
 	return 0;
 }
